feat: Add @nuxtjs/seo module for improved SEO functionality
The commit adds the "@nuxtjs/seo" module to the project's dependencies in the "package.json" file. This module provides enhanced SEO functionality for the website, allowing better control over meta tags, social media sharing, and search engine optimization.

Code changes include the addition of the "@nuxtjs/seo" package to the list of installed modules in the Nuxt configuration file "nuxt.config.ts". This ensures that the module is loaded and configured during the application's initialization.

Additionally, the commit includes updates to the "index.vue" file in the "pages" directory. The useSeoMeta function is utilized to set the title, Twitter card, and Open Graph (OG) properties for the home page. This helps improve the website's visibility and presentation in search engine results and social media platforms.

The commit also introduces a new component file "SpeakerImage.vue" in the "components/OgImage" directory. This component defines the structure and content of the speaker image used for Open Graph (OG) images. It includes a background gradient, logo, title, and position of the speaker.

Lastly, the commit includes minor style updates to the "Speakers.vue" file in the "components/Home" directory. The background color of the pagination bullet in the Swiper component is changed to a green shade.

These changes aim to enhance the website's SEO capabilities, improve the visual presentation of speaker images in social media previews, and provide a better user experience for visitors.